backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
PyMuPDF==1.24.1
python-dotenv==1.0.1
google-generativeai==0.7.1
Flask-Cors==4.0.1